Time: 23 ms (22.44%), Space: 48 MB (43.47%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.java
@@ -7,7 +7,7 @@ class Solution {
         }
         while (left < right) {
             int mid = left + (right - left) / 2;
-            int hoursNeeded = 0; 
+            long hoursNeeded = 0; 
             for (int pile : piles) {
                 hoursNeeded += (int) Math.ceil((double) pile / mid); 
             }
